Implemented and briefly tested: game rules in updatemap endpoint, checks for sufficient funds and meets numOfNodeType criterias.
settled on db schema, polymorphic data structures such as nodes, equipment and character attributes are to be serialised into strings, and stored as jsonb.

requires serialisation without custom SerialisationSupport class, and deserialisation into these polymorphic data structures requires custom SerialisationSupport class. Extends newtonsoft's ReadJson.

+++ MM_API/Database/Postgres/DbSchema/t_Treasury.cs
@@ -16,7 +16,7 @@ namespace MM_API.Database.Postgres.DbSchema
         [ForeignKey("fk_user_id")]
         public t_User User { get; set; }
 
-        public BigInteger treasury_coin { get; set; }
+        public int treasury_coin { get; set; }
 
         public double treasury_gainrate { get; set; }
